ramdisk/zips: updates and ThunderTweaks v1.0.4.2

+++ builds/zip-p/common/ts/ts_clean.sh
@@ -3,13 +3,13 @@
 # Clean app in packages.xml
 # Remove apk in data/app, data/data, and system
 #
-# v3.0.4
+# v3.0.5
 # By morogoku 
 # http://www.espdroids.com
 #
 #
 # Usage:
-#	ts_clean3.sh [apk_package] -[options]
+#	ts_clean.sh [apk_package] -[options]
 #
 # Options:
 #	a: remove apk in /data/app/apk_package
@@ -18,35 +18,24 @@
 # 
 # example:
 #	clean only in packages.xml:
-#	    ts_clean3.sh com.sec.android.app.music
+#	    ts_clean.sh com.sec.android.app.music
 #
 #	clean in packages.xml and remove apk in data/app:
-#	    ts_clean3.sh com.sec.android.app.music -a
+#	    ts_clean.sh com.sec.android.app.music -a
 #
 #	clean on packages.xml, remove in data/app, remove data, and remove in system
-#	    ts_clean3.sh com.sec.android.app.music -ads
+#	    ts_clean.sh com.sec.android.app.music -ads
 #
 #
 # In a updater-script file, copy ts_clean.sh to tmp and:
-#   run_program("/tmp/ts_clean3.sh", "com.sec.android.app.music", "-ads");
+#   run_program("/tmp/ts_clean.sh", "com.sec.android.app.music", "-ads");
 #
 #
 
 
 
 # Busybox 
-BB=/sbin/busybox
-
-
-# Prepare packages.xml
-cd /data/system
-
-if [ -f packages-bak.xml ]; then
-	rm -f packages-bak.xml
-fi
-
-# Backup packages.xml
-cp packages.xml packages-bak.xml
+BB=/sbin/busybox;
 
 
 run_script(){
@@ -99,45 +88,63 @@ run_script(){
 }
 
 
-# If package is entered run script
-if [[ ! -z $1 ]]; then
+# If packages.xml not exist, abort
+if [ -f /data/system/packages.xml ]; then
 
-	ruta="name=\"$1\"";
-	pkg=$1;
-	$BB echo "MC3[i]: Searching $1 in packages.xml...";
-	run_script;
+	# Prepare packages.xml
+	cd /data/system
 
-	# If entered options
-	if [[ ! -z $2 ]] && [[ $xfound = 1 ]]; then
-	
-		# If option "a" Remove apk from /data/app
-		case $2 in *a*)				
-			rm -rf /data/app/$pkg*
-			$BB echo "MC3[i]: Option a -> Removing from /data/app ..."
-		;; esac
-		
-		# If option "d" Remove apk data from /data/data
-		case $2 in *d*)	
-			rm -rf /data/data/$pkg;
-			$BB echo "MC3[i]: Option d -> Removing from /data/data ...";
-		;; esac
-		
-		# If option "s" Remove apk from system if exist
-		case $2 in *s*)	
-			if [ "$system_path" != "" ]; then
-				rm -rf $system_path;
-				$BB echo "MC3[i]: Option s -> Removing $system_path ...";
-			else
-				$BB echo "MC3[i]: Option s -> Not found apk in system";
-			fi
-		;; esac
-
-	else
-		if [[ $xfound = 1 ]]; then
-			$BB echo "MC3[i]: No options entered";
-		fi
+	if [ -f packages-bak.xml ]; then
+		rm -f packages-bak.xml
 	fi
 
+	# Backup packages.xml
+	cp packages.xml packages-bak.xml
+	
+
+	# If package is entered run script
+	if [[ ! -z $1 ]]; then
+
+		ruta="name=\"$1\"";
+		pkg=$1;
+		$BB echo "MC3[i]: Searching $1 in packages.xml...";
+		run_script;
+
+		# If entered options
+		if [[ ! -z $2 ]] && [[ $xfound = 1 ]]; then
+		
+			# If option "a" Remove apk from /data/app
+			case $2 in *a*)				
+				rm -rf /data/app/$pkg*
+				$BB echo "MC3[i]: Option a -> Removing from /data/app ..."
+			;; esac
+			
+			# If option "d" Remove apk data from /data/data
+			case $2 in *d*)	
+				rm -rf /data/data/$pkg;
+				$BB echo "MC3[i]: Option d -> Removing from /data/data ...";
+			;; esac
+			
+			# If option "s" Remove apk from system if exist
+			case $2 in *s*)	
+				if [ "$system_path" != "" ]; then
+					rm -rf $system_path;
+					$BB echo "MC3[i]: Option s -> Removing $system_path ...";
+				else
+					$BB echo "MC3[i]: Option s -> Not found apk in system";
+				fi
+			;; esac
+
+		else
+			if [[ $xfound = 1 ]]; then
+				$BB echo "MC3[i]: No options entered";
+			fi
+		fi
+
+	else
+		$BB echo "MC3[E]: Missing package name"
+	fi
+	
 else
-	$BB echo "MC3[E]: Missing package name"
+	$BB echo "MC3[E]: /data wiped, packages.xml not exist, abort"
 fi
